for remote repository

+++ ScriptEngine/Engine/src/Helpers/casts.h
@@ -13,11 +13,11 @@ public:
 	{
 		glm::vec3 res(0.f);
 
-		float scaleX = RenderEngine::Renderer::ViewportSize.x / cellSize.x;
+		float scaleX = static_cast<float>(RenderEngine::Renderer::ViewportSize.x / cellSize.x);
 		res.x = ivec2.x * scaleX;
 		//res.x = std::max(res.x, (float)((RenderEngine::Renderer::ViewportSize.x - RenderEngine::Renderer::ViewportOffset.x) / cellSize.x));
 
-		float scaleY = RenderEngine::Renderer::ViewportSize.y / cellSize.y;
+		float scaleY = static_cast<float>(RenderEngine::Renderer::ViewportSize.y / cellSize.y);
 		res.y = ivec2.y * scaleY;
 		//res.y = std::max(res.y, (float)((RenderEngine::Renderer::ViewportSize.y - RenderEngine::Renderer::ViewportOffset.y) / cellSize.y));
